refactor: useBundleDocuments only invoked when release menu is opened
